+++ src/tap/control/hold_repeat_command_mapping.cpp
@@ -43,7 +43,7 @@ void HoldRepeatCommandMapping::executeCommandMapping(const RemoteMapState &currS
     {
         // While Commands may not be scheduled this prevents the unnecessary call of the
         // removeCommand function from the scheduler.
-        if (commandsScheduled)
+        if (commandsScheduled && endCommandsWhenNotHeld)
         {
             removeCommands();
             commandsScheduled = false;